+++ include/system/info.h
@@ -20,12 +20,12 @@ public:
     // Modifications to this map requires adjustments at MKBI and SETUP
     struct Boot_Map
     {
-        volatile unsigned long n_cpus;    // Number of CPUs in SMPs
+        volatile unsigned short n_cpus;   // Number of CPUs in SMPs
         PAddr mem_base;                   // Memory base address
         PAddr mem_top;                    // Memory top address
         PAddr mio_base;                   // Memory-mapped I/O base address
         PAddr mio_top;                    // Memory-mapped I/O top address
-        int node_id;                      // Local node id in SAN (-1 => RARP)
+        short node_id;                    // Local node id in SAN (-1 => RARP)
         int space_x;                      // Spatial coordinates of a node (-1 => mobile)
         int space_y;                      //
         int space_z;                      //
